requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
reportlab==4.0.4
openpyxl==3.1.2
PyPDF2==3.0.1
bcrypt==4.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.7
requests==2.31.0
email-validator==2.0.0
Pillow==10.0.0

